Refactored code to comply with Rubocop rules
