+++ Gemfile
@@ -1,10 +1,10 @@
 # frozen_string_literal: true
 
-source "https://rubygems.org"
+source 'https://rubygems.org'
 
-git_source(:github) {|repo_name| "https://github.com/#{repo_name}" }
+git_source(:github) { |repo_name| "https://github.com/#{repo_name}" }
 
-gem "bundler"
-gem "prawn"
-gem "prawn-svg"
-gem "rubocop", require: false
+gem 'bundler'
+gem 'prawn'
+gem 'prawn-svg'
+gem 'rubocop', require: false
